% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSMSSRSS.R
\name{calculate_sums_of_squares}
\alias{calculate_sums_of_squares}
\title{Title Calculate TSS, MSS, RSS}
\usage{
calculate_sums_of_squares(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector representing the actual observed values (e.g., the Height of trees)}

\item{predicted}{A numeric vector representing the predicted values (e.g., from a linear model)}
}
\value{
A list containing TSS, MSS, and RSS
}
\description{
Title Calculate TSS, MSS, RSS
}
\examples{
actual <- c(10, 15, 20, 25, 30)
predicted <- c(12, 16, 19, 24, 28)
calculate_sums_of_squares(actual, predicted)
}
