% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREEEE.R, R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM Dataset}
\format{
## `fire`
A data frame with X rows and Y columns:
\describe{
  \item{Distance}{The distance from the fire location (in relevant units)}
  \item{Damage}{The estimated damage caused by the fire (in relevant units)}
}

## `fire`
A data frame with X rows and Y columns:
\describe{
  \item{Distance}{The distance from the fire location (in relevant units)}
  \item{Damage}{The estimated damage caused by the fire (in relevant units)}
}
}
\source{
FIREDAM internal report (data obtained from `FIREDAM.csv`)

FIREDAM internal report (data obtained from `FIREDAM.csv`)
}
\usage{
fire

fire
}
\description{
A dataset containing fire damage statistics from the FIREDAM report.

A dataset containing fire damage statistics from the FIREDAM report.
}
\keyword{datasets}
