% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval and Histogram Plot}
\usage{
myboot2(
  iter = 10000,
  x,
  fun = "mean",
  alpha = 0.05,
  cx = 1.5,
  xlab = "Statistic",
  ylab = "Density",
  ...
)
}
\arguments{
\item{iter}{An integer specifying the number of bootstrap iterations. Default is 10000.}

\item{x}{A numeric vector containing the sample data to bootstrap.}

\item{fun}{A character string or function for the statistic to compute for each bootstrap sample. Default is "mean".}

\item{alpha}{A numeric value for the significance level. Default is 0.05 (for a 95\% confidence interval).}

\item{cx}{A numeric value specifying the scaling factor for text size in confidence interval annotations. Default is 1.5.}

\item{xlab}{A character string for the label of the x-axis in the histogram. Default is "Statistic".}

\item{ylab}{A character string for the label of the y-axis in the histogram. Default is "Density".}

\item{...}{Additional graphical parameters to be passed to \code{hist()}, such as \code{col} for color.}
}
\description{
Bootstrap Confidence Interval and Histogram Plot
}
\examples{
# Example usage of myboot2 function
set.seed(39)
sam <- rnorm(25, mean = 25, sd = 10)
myboot2(iter = 10000, x = sam, fun = "mean", alpha = 0.05, xlab = "Mean", ylab = "Density", col = "Purple", cx = 1.5)

}
