% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem for Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 4, ...)
}
\arguments{
\item{n}{Sample size for each iteration}

\item{iter}{Number of iterations}

\item{lambda}{Poisson parameter}

\item{...}{Additional graphical parameters}
}
\value{
A histogram of sample means and a theoretical normal curve
}
\description{
This function generates a sample from a Poisson distribution, calculates sample means,
and plots the histogram of the sample means along with a theoretical normal curve.
}
