% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{mycurve}
\alias{mycurve}
\title{mycurve}
\usage{
mycurve(mu, sigma, a)
}
\arguments{
\item{mu}{Numeric, the mean of the normal distribution.}

\item{sigma}{Numeric, the standard deviation of the normal distribution.}

\item{a}{Numeric, the value up to which the area under the curve is shaded.}
}
\value{
A list containing the mean (`mu`), the standard deviation (`sigma`),
and the calculated cumulative probability (`probability`).
}
\description{
mycurve
}
\examples{
mycurve(mu = 10, sigma = 5, a = 6)
}
