% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlike.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A log likelihood function.}

\item{x}{A vector of observed data values.}

\item{param}{A vector of possible parameter values.}

\item{...}{Additional arguments for plot customization.}
}
\value{
A list containing the index, parameter estimate, and log likelihood value at the maximum.
}
\description{
This function calculates the maximum likelihood estimate for a parameter using a specified log likelihood function.
}
\examples{
data <- c(3, 3, 4, 5)
mymaxlik(logbin, data, seq(0, 1, length = 1000), xlab = expression(pi), main = "MLE for Binomial")
}
