% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Calculate and Plot p-Value for t-Statistic}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The calculated t-statistic.}

\item{xmax}{The maximum x-axis limit for the plot. Default is 4.}

\item{n}{The sample size. Default is 20.}

\item{alpha}{The significance level for the plot. Default is 0.05.}
}
\value{
A list containing the critical value (q) and the p-value.
}
\description{
This function calculates the p-value for a given t-statistic and plots the
area under the t-distribution curve corresponding to the p-value.
}
\examples{
# Calculate and plot p-value for t0 = 2 with sample size n = 30
Yourpackage::mypvalue(t0 = 2, n = 30)

}
