% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Number of Tickets to Sell}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Total number of tickets available}

\item{gamma}{The risk level (probability of exceeding the available seats)}

\item{p}{The probability of a ticket holder showing up}
}
\value{
A list containing the calculated number of tickets for the discrete case (`nd`)
and for the continuous case (`nc`), as well as the input parameters.
}
\description{
This function calculates the optimal number of tickets to sell based on
a binomial distribution (discrete case) and its normal approximation (continuous case).
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
