% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{simulate_binomial}
\alias{simulate_binomial}
\title{Title: Simulate Binomial Trials}
\usage{
simulate_binomial(n_trials = 10, prob_success = 0.5, n_experiments = 100)
}
\arguments{
\item{n_trials}{- integer, number of trials in each experiment}

\item{prob_success}{- double, probability of success in each trial}

\item{n_experiments}{- integer, number of experiments to run}
}
\value{
A data frame with the results of each experiment
}
\description{
Title: Simulate Binomial Trials
}
\examples{
simulate_binomial() 
simulate_binomial(10, 0.5, 100)
}
